take displayProductImages into account when rendering the products in the searchbox
