+++ projects/storefrontlib/src/cms-components/navigation/search-box/search-box.component.spec.ts
@@ -47,7 +47,7 @@ export class MockMediaComponent {
   @Input() alt;
 }
 
-describe('SearchBoxComponent', () => {
+fdescribe('SearchBoxComponent', () => {
   let searchBoxComponent: SearchBoxComponent;
   let fixture: ComponentFixture<SearchBoxComponent>;
   let serviceSpy: SearchBoxComponentService;
