+++ projects/storefrontlib/src/cms-components/navigation/search-box/search-box.model.ts
@@ -1,6 +1,7 @@
 export interface SearchBoxConfig {
   displaySuggestions: boolean;
   displayProducts: boolean;
+  displayProductImages?: boolean;
   maxProducts?: number;
   maxSuggestions?: number;
   minCharactersBeforeRequest?: number;
